Add files via upload
Comit Final

+++ src/core/pet/Parrot.java
@@ -14,9 +14,11 @@ import java.util.ArrayList;
  */
 public class Parrot extends Pet {
 
-    public Parrot(String name, int age, float height, float weight, Owner owner) {
+    public Parrot(String name, int age, double height, double weight, Owner owner) {
         super(name, age, height, weight, owner);
     }
+
+
